requirements.txt
psycopg2-binary
streamlit 
streamlit_authenticator